+++ examples/structs.rs
@@ -4,7 +4,7 @@
 //  Created:
 //    26 Dec 2024, 12:12:07
 //  Last edited:
-//    09 Jan 2025, 20:24:23
+//    04 Feb 2025, 16:30:20
 //  Auto updated?
 //    Yes
 //
@@ -15,7 +15,7 @@
 use std::hash::{DefaultHasher, Hasher as _};
 use std::marker::PhantomData;
 
-use better_derive::{Debug, Eq, Hash, PartialEq};
+use better_derive::{Clone, Debug, Eq, Hash, PartialEq};
 
 
 /***** HELPER FUNCTIONS *****/
@@ -32,15 +32,15 @@ fn hash<T: std::hash::Hash>(obj: T) -> u64 {
 
 /***** EXAMPLES *****/
 /// Example unit struct as usual.
-#[derive(Debug, Eq, Hash, PartialEq)]
+#[derive(Clone, Debug, Eq, Hash, PartialEq)]
 struct Foo;
 
 /// Example tuple struct as usual.
-#[derive(Debug, Eq, Hash, PartialEq)]
+#[derive(Clone, Debug, Eq, Hash, PartialEq)]
 struct Bar((), bool, String);
 
 /// Example struct struct as usual.
-#[derive(Debug, Eq, Hash, PartialEq)]
+#[derive(Clone, Debug, Eq, Hash, PartialEq)]
 struct Baz {
     a: (),
     b: bool,
@@ -52,7 +52,7 @@ struct Baz {
 struct DontImplementAnything;
 
 /// Special struct with generics that don't have to be debug.
-#[derive(Debug, Eq, Hash, PartialEq)]
+#[derive(Clone, Debug, Eq, Hash, PartialEq)]
 struct PhantomStruct<T> {
     _f: PhantomData<T>,
 }
@@ -63,6 +63,13 @@ struct PhantomStruct<T> {
 
 /***** ENTRYPOINT *****/
 fn main() {
+    assert_eq!(Foo.clone(), Foo);
+    assert_eq!(Bar((), true, "Hello, world!".into()).clone(), Bar((), true, "Hello, world!".into()));
+    assert_eq!(Baz { a: (), b: true, c: "Hello, world!".into() }.clone(), Baz { a: (), b: true, c: "Hello, world!".into() });
+    assert_eq!(PhantomStruct::<DontImplementAnything> { _f: PhantomData }.clone(), PhantomStruct::<DontImplementAnything> { _f: PhantomData });
+
+
+
     assert_eq!(format!("{:?}", Foo), "Foo");
     assert_eq!(format!("{:#?}", Foo), "Foo");
     assert_eq!(format!("{:?}", Bar((), true, "Hello, world!".into())), "Bar((), true, \"Hello, world!\")");