+++ CHANGELOG.md
@@ -4,6 +4,14 @@ This file keeps track of notable changes to the `better-derive`-crate.
 The project uses [semantic versioning](https://semver.org). As such, breaking changes are indicated by **(BREAKING)**.
 
 
+## v1.1.0 - 2025-02-04
+### Added
+- The `Clone` derive macro.
+
+### Changed
+- Updated `proc-macro2` and `syn` dependencies to latest versions.
+
+
 ## v1.0.1 - 2025-01-09
 ### Added
 - `Eq`, `Hash` and `PartialEq` to the examples.